+++ app/Data_Definitions/Data.py
@@ -2,47 +2,102 @@
 ## Data
 ## ====================
 
-## squareFootage is float
-squareFootage = float
-## interp. square footage of a property
 
-sf  = 245.4
+class SquareFootage:
+    ## SquareFootage is float
+    ## interp. square footage of a property
+        
+    ## sf1  = SquareFootage(245.4)
+        
+    '''def fn_for_SquareFootage(sf):
+            # ... sf
+            pass
 
-def fn_for_squareFootage(squareFootage):
-    # ... squareFootage
-    pass
+        ## Template rules used:
+        ##  - atomic non-distinct:  float (Number)'''
+    
+    def __init__(self, squarefootage):
+        self.set_squarefootage(squarefootage)    
+    
+    def set_squarefootage(self, squarefootage):
+        
+        try:
+            if isinstance(squarefootage, str) or isinstance(squarefootage, bool): # string or boolean
+                raise TypeError
+            if squarefootage < 0 : # neg
+                raise ValueError    
+        except TypeError: # loggin the error, notifying the user, reverting to a defult value or other appropriate actions 
+            raise TypeError
+        except ValueError:
+            raise ValueError 
+        else:
+            self.value = squarefootage
 
-## Template rules used:
-##  - atomic non-distinct:  float (Number)
 
 ## ====================
 
-## numofBedrooms is Integer[0,10]
-numofBedrooms = Integer[0,10]
-## interp. The number of bedrooms in a property
+class NumofBedrooms:
+    ## numofBedrooms is Integer[0,10]    
+    ## interp. The number of bedrooms in a property
 
-nb = 4
+    ## nb1 = NumofBedrooms(4)
 
-def fn_for_numofBedrooms(numofBedrooms):
-    # ... numofBedrooms
-    pass
+    '''
+    def fn_for_numofBedrooms(numofBedrooms):
+        # ... numofBedrooms
+        pass
 
-## Template rules used:
-##  - atomic non-distinct:  Integer[0,10]
+    ## Template rules used:
+    ##  - atomic non-distinct:  Integer[0,10]'''
 
+    def __init__(self, numofbedrooms):
+        self.set_numofBedrooms(numofbedrooms)
+
+    def set_numofBedrooms(self, numofBedrooms):
+            
+            try:
+                if isinstance(numofBedrooms, str) or isinstance(numofBedrooms, bool) or isinstance(numofBedrooms, float): # string or boolean
+                    raise TypeError
+                if numofBedrooms < 0 or numofBedrooms > 10  : # neg
+                    raise ValueError    
+            except TypeError: # loggin the error, notifying the user, reverting to a defult value or other appropriate actions 
+                raise TypeError
+            except ValueError:
+                raise ValueError 
+            else:
+                self.value = numofBedrooms
 ## ====================
 
-## numofBathrooms is Integer[1,10] 
-## interp. The number of bathrooms in a property
+class NumofBathrooms:
+    ## numofBathrooms is Integer[1,10] 
+    ## interp. The number of bathrooms in a property
 
-nb = 2
+    ## nb1 = NumofBathrooms(2)
 
-def fn_for_numofBathrooms(numofBathrooms):
-    ## ... numofBathrooms
-    pass
+    '''
+    def fn_for_numofBathrooms(numofBathrooms):
+        ## ... numofBathrooms
+        pass
 
-## Template rules used:
-##  - atomic non-distinct:  Integer[0,10]
+    ## Template rules used:
+    ##  - atomic non-distinct:  Integer[0,10]'''
+
+    def __init__(self, numofbathrooms):
+        self.set_numofBathrooms(numofbathrooms)
+
+    def set_numofBathrooms(self, numofBathrooms):
+            
+        try:
+            if isinstance(numofBathrooms, str) or isinstance(numofBathrooms, bool) or isinstance(numofBathrooms, float): # string or boolean
+                raise TypeError
+            if numofBathrooms < 1 or numofBathrooms > 10  : # neg and 0 
+                raise ValueError    
+        except TypeError: # loggin the error, notifying the user, reverting to a defult value or other appropriate actions 
+            raise TypeError
+        except ValueError:
+            raise ValueError 
+        else:
+            self.value = numofBathrooms
